working on quadrature operations for clenshaw curtis grids

+++ datadriven/python/uq/learner/SimulationLearner.py
@@ -270,7 +270,7 @@ class SimulationLearner(Learner):
         Create a set of all collocation nodes
         """
         gs = self.grid.getStorage()
-        ps = np.ndarray([gs.getSize(), gs.getDimension()], dtype='float32')
+        ps = np.ndarray([gs.getSize(), gs.getDimension()], dtype='float')
         p = DataVector(gs.getDimension())
         for i in xrange(gs.getSize()):
             gs.getCoordinates(gs.getPoint(i), p)
@@ -311,7 +311,7 @@ class SimulationLearner(Learner):
         # parse them to a numpy array
         gs = self.grid.getStorage()
         p = DataVector(gs.getDimension())
-        ans = np.ndarray([len(newCollocationNodes), gs.getDimension()], dtype='float32')
+        ans = np.ndarray([len(newCollocationNodes), gs.getDimension()], dtype='float')
         for i, gp in enumerate(newCollocationNodes):
             gs.getCoordinates(gp, p)
             ans[i, :] = p.array()
